[SCD-64] rename cest
